service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
tenacity==9.0.0
orjson==3.10.12
python-json-logger==3.2.0
prometheus-client==0.21.1
tiktoken==0.8.0

opensearch-py==2.7.1
qdrant-client==1.12.2

opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-httpx==0.50b0

